fix error in mesh create from h5m file

+++ tests/test_r2s.py
@@ -390,7 +390,7 @@ def test_irradiation_setup_unstructured_nondef_tag():
         i += 1
 
 
-def _r2s_test_step1(r2s_run_dir):
+def _r2s_test_step1(r2s_run_dir, remove_step1_out=True):
     os.chdir(thisdir)
     # copy ../scripts/r2s.py to r2s_run_dir/r2s.py
     os.chdir("..")
@@ -419,10 +419,13 @@ def _r2s_test_step1(r2s_run_dir):
     f3 = filecmp.cmp(alara_fluxin, exp_alara_fluxin)
 
     # remove test output files
-    os.remove(alara_inp)
+    if remove_step1_out:
+        # these files are used in step2
+        os.remove(blank_mesh)
+        os.remove(alara_inp)
+    # these files are not used in step2
     os.remove(alara_fluxin)
     os.remove(alara_matlib)
-    os.remove(blank_mesh)
     os.remove(step1_file)
     os.remove(dst)
 
@@ -431,7 +434,7 @@ def _r2s_test_step1(r2s_run_dir):
     assert_equal(f3, True)
 
 
-def _r2s_test_step2(r2s_run_dir):
+def _r2s_test_step2(r2s_run_dir, remove_step1_out=True):
     os.chdir(thisdir)
     # copy ../scripts/r2s.py to r2s_run_dir/r2s.py
     os.chdir("..")
@@ -441,9 +444,10 @@ def _r2s_test_step2(r2s_run_dir):
 
     # output files of r2s step1
     alara_inp = os.path.join(r2s_run_dir, "alara_inp")
-    copyfile(os.path.join(r2s_run_dir, "exp_alara_inp"), alara_inp)
     blank_mesh = os.path.join(r2s_run_dir, "blank_mesh.h5m")
-    copyfile(os.path.join(r2s_run_dir, "exp_blank_mesh.h5m"), blank_mesh)
+    if remove_step1_out:
+        copyfile(os.path.join(r2s_run_dir, "exp_alara_inp"), alara_inp)
+        copyfile(os.path.join(r2s_run_dir, "exp_blank_mesh.h5m"), blank_mesh)
 
     # run r2s step2
     os.chdir(r2s_run_dir)
@@ -476,7 +480,7 @@ def _r2s_test_step2(r2s_run_dir):
     assert_equal(f5, True)
 
 
-def test_r2s_script():
+def test_r2s_script_step_by_step():
 
     # skip test without dagmc
     try:
@@ -484,17 +488,43 @@ def test_r2s_script():
     except ImportError:
         raise SkipTest
 
+    remove_step1_out = True
     r2s_run_dir = os.path.join(
         thisdir, "files_test_r2s", "r2s_examples", "r2s_run")
-    _r2s_test_step1(r2s_run_dir)
-    _r2s_test_step2(r2s_run_dir)
+    _r2s_test_step1(r2s_run_dir, remove_step1_out)
+    _r2s_test_step2(r2s_run_dir, remove_step1_out)
     # test sub-voxel r2s
     r2s_run_dir = os.path.join(
         thisdir, "files_test_r2s", "r2s_examples", "subvoxel_r2s_run")
-    _r2s_test_step1(r2s_run_dir)
-    _r2s_test_step2(r2s_run_dir)
+    _r2s_test_step1(r2s_run_dir, remove_step1_out)
+    _r2s_test_step2(r2s_run_dir, remove_step1_out)
     # test unstructured r2s
     r2s_run_dir = os.path.join(
         thisdir, "files_test_r2s", "r2s_examples", "unstructured_r2s_run")
-    _r2s_test_step1(r2s_run_dir)
-    _r2s_test_step2(r2s_run_dir)
+    _r2s_test_step1(r2s_run_dir, remove_step1_out)
+    _r2s_test_step2(r2s_run_dir, remove_step1_out)
+
+
+def test_r2s_script():
+
+   # skip test without dagmc
+   try:
+       from pyne import dagmc
+   except ImportError:
+       raise SkipTest
+
+   remove_step1_out = False
+   r2s_run_dir = os.path.join(
+       thisdir, "files_test_r2s", "r2s_examples", "r2s_run")
+   _r2s_test_step1(r2s_run_dir, remove_step1_out)
+   _r2s_test_step2(r2s_run_dir, remove_step1_out)
+   # test sub-voxel r2s
+   r2s_run_dir = os.path.join(
+       thisdir, "files_test_r2s", "r2s_examples", "subvoxel_r2s_run")
+   _r2s_test_step1(r2s_run_dir, remove_step1_out)
+   _r2s_test_step2(r2s_run_dir, remove_step1_out)
+   # test unstructured r2s
+   r2s_run_dir = os.path.join(
+       thisdir, "files_test_r2s", "r2s_examples", "unstructured_r2s_run")
+   _r2s_test_step1(r2s_run_dir, remove_step1_out)
+   _r2s_test_step2(r2s_run_dir, remove_step1_out)
